+++ Source/Core/GlobalVariables.h
@@ -2,13 +2,13 @@
 
 
 namespace Global {
-	extern int windowWidth;
-	extern int windowHeight;
-	extern unsigned long long int frameCounter;
-	extern bool status;
-	extern short frameDelay;
-	extern unsigned int TickTimer;
-	extern unsigned int LogicTimeEnd;
-	extern unsigned int RenderingTimeEnd;
-	extern unsigned int inputDelay;
+	inline int windowWidth = 1400;
+	inline int windowHeight = 800;
+
+	inline unsigned long long int logicCount = 0;
+	inline unsigned long long int frameCount = 0;
+	inline bool running = true;
+
+	inline unsigned int logicDelay = 16; // ms
+	inline unsigned int frameDelay = 16; // ms
 }